functions/requirements.txt
firebase-functions==0.4.2
firebase-admin==6.2.0
flask==2.3.2
google-cloud-firestore==2.11.1
openai==1.3.0
stripe
tiktoken
groq